requirements.txt
streamlit>=1.28.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
python-dotenv>=1.0.0
openai>=1.0.0
plotly>=5.0.0
faiss-cpu>=1.7.0
